Add extra_hosts to docker agent
