+++ src/prefect/cli/server.py
@@ -11,20 +11,7 @@ import yaml
 
 from prefect import config
 from prefect.utilities.configuration import set_temporary_config
-
-
-def platform_is_linux():
-    return sys.platform.startswith("linux")
-
-
-def get_docker_ip():
-    """Get local docker internal IP without using shell=True in subprocess"""
-    from subprocess import Popen, PIPE
-
-    ip_route_proc = Popen(["ip", "route"], stdout=PIPE)
-    grep_proc = Popen(["grep", "docker0"], stdin=ip_route_proc.stdout, stdout=PIPE)
-    awk_proc = Popen(["awk", "{print $9}"], stdin=grep_proc.stdout, stdout=PIPE)
-    return awk_proc.communicate()[0].strip().decode()
+from prefect.utilities.docker_util import platform_is_linux, get_docker_ip
 
 
 def make_env(fname=None):
